+++ dfm_tools/__init__.py
@@ -6,7 +6,7 @@ __author__ = """Jelmer Veenstra"""
 __email__ = "jelmer.veenstra@deltares.nl"
 __version__ = "0.33.1"
 
-from dfm_tools.deprecated import *
+from dfm_tools.deprecated_functions import *
 from dfm_tools.download import *
 from dfm_tools.get_nc import *
 from dfm_tools.get_nc_helpers import *